requirements.txt
flask
scikit-learn
numpy
pillow
h5py
